move common functions into separate utils files
