+++ data/utils/getCSVHeaders.js
@@ -1,0 +1,5 @@
+const getHeaders = (string) => {
+	return string.split(",")
+}
+
+module.exports = getHeaders;
